# docs/API_SPEC.yaml
openapi: 3.0.3
info:
  title: Runeterra Lore API
  version: 1.0.0
  description: >
    Canonical API for champions, skins, and regions.
    Data is sourced from Riot's Data Dragon (ddragon) with local fallbacks.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://runeterra-lore.riotgames.components
    description: Production

paths:
  /champions:
    get:
      summary: List champions
      description: Return all champions, optionally filtered by name substring.
      parameters:
        - name: search
          in: query
          description: Case-insensitive substring filter on champion name
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of champions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChampionSummary"

  /champions/{champion_id}:
    get:
      summary: Get champion detail
      description: Return detailed information about a single champion.
      parameters:
        - name: champion_id
          in: paths
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Champion detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChampionDetail"
        "404":
          description: Champion not found

  /skins:
    get:
      summary: List skins
      description: Return all skins normalized from champion data.
      responses:
        "200":
          description: List of skins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skin"

  /skins/{skin_id}:
    get:
      summary: Get skin detail
      description: Return detailed information about a single skin.
      parameters:
        - name: skin_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Skin detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skin"
        "404":
          description: Skin not found

  /regions:
    get:
      summary: List regions
      description: Return all regions, optionally filtered by name substring.
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of regions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RegionSummary"

  /regions/{region_id}:
    get:
      summary: Get region detail
      description: Return detailed information about a single region.
      parameters:
        - name: region_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Region detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegionDetail"
        "404":
          description: Region not found

components:
  schemas:
    ChampionSummary:
      type: object
      properties:
        id:
          type: string
          example: Ahri
        key:
          type: string
          example: "103"
        name:
          type: string
          example: Ahri
        title:
          type: string
          example: the Nine-Tailed Fox
        tags:
          type: array
          items:
            type: string
          example: ["Mage", "Assassin"]
        icon:
          type: string
          example: /cdn/14.18.1/img/champion/Ahri.png

    ChampionDetail:
      allOf:
        - $ref: "#/components/schemas/ChampionSummary"
        - type: object
          properties:
            lore:
              type: string
              example: "Innately connected to the magic of the spirit realm..."
            info:
              type: object
              properties:
                attack: { type: integer, example: 3 }
                defense: { type: integer, example: 4 }
                magic: { type: integer, example: 8 }
                difficulty: { type: integer, example: 5 }
            stats:
              type: object
              additionalProperties: true
              description: Full numerical stats block
            abilities:
              type: object
              description: Mapping of ability key → name
              additionalProperties:
                type: string
              example:
                Q: Orb of Deception
                W: Fox-Fire
                E: Charm
                R: Spirit Rush
            skins:
              type: array
              items:
                $ref: "#/components/schemas/ChampionSkin"

    ChampionSkin:
      type: object
      properties:
        id:
          type: string
          example: ahri_1
        num:
          type: integer
          example: 1
        name:
          type: string
          example: Dynasty Ahri
        chromas:
          type: boolean
          example: false
        splash:
          type: string
          example: /cdn/img/champion/splash/Ahri_1.jpg
        loading:
          type: string
          example: /cdn/img/champion/loading/Ahri_1.jpg

    Skin:
      type: object
      properties:
        id:
          type: string
          example: karma_1
        championId:
          type: string
          example: Karma
        name:
          type: string
          example: Order of the Lotus Karma
        num:
          type: integer
          example: 3
        chromas:
          type: boolean
          example: false
        splash:
          type: string
          example: /cdn/img/champion/splash/Karma_3.jpg
        loading:
          type: string
          example: /cdn/img/champion/loading/Karma_3.jpg
        releaseDate:
          type: string
          format: date
          example: 2017-09-25

    RegionSummary:
      type: object
      properties:
        id:
          type: string
          example: ionia
        name:
          type: string
          example: Ionia
        icon:
          type: string
          example: /assets/regions/ionia.png

    RegionDetail:
      allOf:
        - $ref: "#/components/schemas/RegionSummary"
        - type: object
          properties:
            description:
              type: string
              example: A land of spiritual harmony and martial tradition.
            champions:
              type: array
              items:
                type: string
              example: ["Ahri", "Karma", "Irelia"]
            emblem:
              type: string
              example: /assets/regions/ionia_emblem.png